XML: Handle `XML` in extension
and make it an optional dependency of the `Serde` package.
