+++ Project.toml
@@ -4,7 +4,6 @@ version = "3.1.0"
 
 [deps]
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
-EzXML = "8f5d6c58-4d21-5cfd-889c-e3ad7ee6a615"
 UUIDs = "cf7118a7-6976-5b1a-9a39-7adc72f591a4"
 
 [weakdeps]
@@ -12,12 +11,14 @@ CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
 JSON = "682c06a0-de6a-54ab-a142-c8b1cf79cde6"
 TOML = "fa267f1f-6049-4f14-aa54-33bafae1ed76"
 YAML = "ddb6d928-2868-570f-bddf-ab3f9cf99eb6"
+EzXML = "8f5d6c58-4d21-5cfd-889c-e3ad7ee6a615"
 
 [extensions]
 SerdeCSVExt = "CSV"
 SerdeJSONExt = "JSON"
 SerdeTOMLExt = "TOML"
 SerdeYAMLExt = "YAML"
+SerdeXMLExt = "EzXML"
 
 [compat]
 CSV = "0.10"
