+++ ext/SerdeXMLExt/Par.jl
@@ -63,41 +63,6 @@ function _xml2dict(node::EzXML.Node; dict_type::Type{D}) where {D<:AbstractDict}
     return xml_dict
 end
 
-"""
-    parse_xml(x::AbstractString; kw...) -> Dict{String,Any}
-    parse_xml(x::Vector{UInt8}; kw...) -> Dict{String,Any}
-
-Parse an XML string `x` (or vector of UInt8) into a dictionary.
-
-## Keyword arguments
-- `dict_type::Type{<:AbstractDict} = Dict`: The type of the dictionary to be returned.
-
-## Examples
-
-```julia-repl
-julia> xml = \"\"\"
-           <book id="bk101">
-              <title>Advanced Julia Programming</title>
-              <authors>
-                  <author lang="en">John Doe</author>
-                  <author lang="es">Juan Pérez</author>
-              </authors>
-              <year>2024</year>
-              <price>49.99</price>
-           </book>
-       \"\"\"
-
-julia> parse_xml(xml)
-Dict{String, Any} with 5 entries:
-  "price"   => Dict{String, Any}("_"=>"49.99")
-  "year"    => Dict{String, Any}("_"=>"2024")
-  "id"      => "bk101"
-  "title"   => Dict{String, Any}("_"=>"Advanced Julia Programming")
-  "authors" => Dict{String, Any}("author"=>Dict{String, Any}[Dict("lang"=>"en", "_"=>"John Doe"), Dict("lang"=>"es", "_"=>"Juan Pérez")])
-```
-"""
-function parse_xml end
-
 function parse_xml(x::S; dict_type::Type{D} = Dict{String,Any}, kw...) where {S<:AbstractString,D<:AbstractDict}
     try
         _xml2dict(x; dict_type = dict_type, kw...)
