+++ ext/SerdeXMLExt/SerdeXMLExt.jl
@@ -1,0 +1,12 @@
+module SerdeXMLExt
+
+include("Par.jl")
+using .ParXml
+
+include("De.jl")
+using .DeXml
+
+include("Ser.jl")
+using .SerXml
+
+end
