+++ src/Par/Par.jl
@@ -2,6 +2,3 @@
 
 include("ParQuery.jl")
 using .ParQuery
-
-include("ParXml.jl")
-using .ParXml
